refactor: move create_stac.py script to /stac folder (#489)
