+++ scripts/stac/imagery/create_stac.py
